+++ CHANGELOG.md
@@ -23,6 +23,10 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ## [Unreleased]
 [Unreleased]: https://github.com/atlassian/virtual-users/compare/release-3.13.2...master
 
+### Added
+- Add a default implementation of `Browser.start` that takes `VirtualUserResult` as a parameter.
+- Add `VerboseChromeBrowser`. Resolves [JPERF-844].
+
 ## [3.13.2] - 2022-04-08
 [3.13.2]: https://github.com/atlassian/virtual-users/compare/release-3.13.1...release-3.13.2